Phase A: pylint 4 件解消 (R1725 / W0613 x2 / C0302)
- R1725: super(DashboardServer, self).shutdown を super().shutdown() 化。
  zero-arg super は inner function 内では `RuntimeError: super(): no arguments`
  となるため、watchdog ループを method (`_watchdog_loop`) に持ち上げて method
  スコープで super() を解決できるようにする。
- W0613: signal handler の signum/frame を _signum/_frame にリネーム。
- C0302: tests/test_dashboard.py が 1096 行 (>1000) で too-many-lines。
  Phase A 関連の test class (TestThreadingServer / TestHealthzEndpoint /
  TestServerJsonLifecycle / TestIdleWatchdog / TestRunIntegration) を
  tests/test_dashboard_live.py に切り出した。共通ヘルパ load_dashboard_module
  は test_dashboard.py から再 import する (pytest が tests/ を sys.path に
  入れるので package 化なしで動く)。

pylint 10.00/10、全テスト 280/280 pass。
